refactor: move RowNetworkItem to common component
